requirements.txt
torch
numpy
pytest
tqdm
hydra-core
omegaconf
transformers
datasets
scikit-learn
tensorboard
evaluate